wyt/requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
appdirs==1.4.3
Appium-Python-Client==0.26
atomicwrites==1.3.0
attrs==19.3.0
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
colorlog==4.0.2
cryptography==2.8
cssselect==1.1.0
Django==2.0.3
et-xmlfile==1.0.1
fake-useragent==0.1.11
Flask==1.1.1
greenlet==0.4.15
har2case==0.3.1
HttpRunner==1.5.8
idna==2.8
importlib-metadata==0.23
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.10.3
lxml==4.4.1
MarkupSafe==1.1.1
more-itertools==7.2.0
Naked==0.1.31
openpyxl==3.0.0
packaging==19.2
parse==1.12.1
pluggy==0.13.1
py==1.8.0
pycparser==2.19
pycryptodome==3.9.4
pyee==6.0.0
PyMySQL==0.9.3
pyparsing==2.4.2
pyppeteer==0.0.25
pyquery==1.4.0
pytest==5.4.2
pytest-metadata==1.8.0
pytest-ordering==0.6
pytest-repeat==0.7.0
pytest-rerunfailures==9.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
requests-html==0.10.0
requests-toolbelt==0.9.1
selenium==2.53.6
shellescape==3.4.1
six==1.12.0
soupsieve==1.9.4
tqdm==4.36.1
urllib3==1.25.6
virtualenv==16.6.1
w3lib==1.21.0
wcwidth==0.1.7
websockets==8.0.2
Werkzeug==0.16.0
xlrd==1.2.0
zipp==0.6.0
